requirements.txt
ultralytics==8.2.45
streamlit==1.37.0
streamlit-option-menu
st-theme
cv2-enumerate-cameras